sis-scraper: optimize quacs-rs bitvecs by pruning unnecessary conflict blocks
This change introduces an optimization that allows quacs-rs to require fewer
uint64_t 'bitvecs' by not considering slots that have either 0 or exactly 1 classes in them.
The reasoning for this is there can never be a conflict in that case, so we're just wasting memory keeping them around.

Testing on Spring 2021 data, this reduces the number of u64s from 10 to 7, a 30% reduction